// src/app/core/api/model/lawOffer.ts
/**
 * Documentation for ads portal
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Location } from './location';

export interface LawOffer {
  id: number;
  userFirstName: string;
  title: string;
  description: string;
  phoneNumber?: string;
  phoneCountryCode?: string;
  modifiedDate?: string;
  location?: Location;
  helpMode: Array<LawOffer.HelpModeEnum>;
  helpKind: Array<LawOffer.HelpKindEnum>;
  language: Array<LawOffer.LanguageEnum>;
  type: LawOffer.TypeEnum;
}
export namespace LawOffer {
  export type HelpModeEnum = 'STATIONARY' | 'WITH_ACCESS' | 'ONLINE' | 'BY_EMAIL' | 'BY_PHONE';
  export const HelpModeEnum = {
    Stationary: 'STATIONARY' as HelpModeEnum,
    WithAccess: 'WITH_ACCESS' as HelpModeEnum,
    Online: 'ONLINE' as HelpModeEnum,
    ByEmail: 'BY_EMAIL' as HelpModeEnum,
    ByPhone: 'BY_PHONE' as HelpModeEnum,
  };
  export type HelpKindEnum = 'LABOUR_LAW' | 'IMMIGRATION_LAW' | 'FAMILY_LAW' | 'TAX_LAW' | 'CIVIL_LAW' | 'OTHER';
  export const HelpKindEnum = {
    LabourLaw: 'LABOUR_LAW' as HelpKindEnum,
    ImmigrationLaw: 'IMMIGRATION_LAW' as HelpKindEnum,
    FamilyLaw: 'FAMILY_LAW' as HelpKindEnum,
    TaxLaw: 'TAX_LAW' as HelpKindEnum,
    CivilLaw: 'CIVIL_LAW' as HelpKindEnum,
    Other: 'OTHER' as HelpKindEnum,
  };
  export type LanguageEnum = 'UA' | 'PL' | 'EN' | 'RU';
  export const LanguageEnum = {
    Ua: 'UA' as LanguageEnum,
    Pl: 'PL' as LanguageEnum,
    En: 'EN' as LanguageEnum,
    Ru: 'RU' as LanguageEnum,
  };
  export type TypeEnum = 'LAW';
  export const TypeEnum = {
    Law: 'LAW' as TypeEnum,
  };
}